auto-browser/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-browser"
version = "0.1.0"
description = "Browser automation CLI tool for configurable site scraping"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=10.0.0",
    "langchain-openai>=0.0.2",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "black"
]

[project.scripts]
auto-browser = "browser_automation.cli:main"

[tool.setuptools]
packages = ["browser_automation"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
